+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Python 3.6 support
+- Python 3.7-dev support (allowed to fail)
 
-*Nothing here yet*
+### Fixed
+- Do not overwrite existing attributes of objects with sure properties (when. should, ...). Refs #124, #128
 
 ## [v1.4.0]
 ### Added