+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+*Nothing to release yet*
+
+## [v1.4.1]
 ### Added
 - Python 3.6 support
 - Python 3.7-dev support (allowed to fail)
@@ -37,7 +41,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Please see `git log`
 
-[Unreleased]: https://github.com/gabrielfalcao/sure/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/gabrielfalcao/sure/compare/v1.4.1...HEAD
+[v1.4.1]: https://github.com/gabrielfalcao/sure/compare/1.4.0...v1.4.1
 [v1.4.0]: https://github.com/gabrielfalcao/sure/compare/1.3.0...v1.4.0
 [v1.3.0]: https://github.com/gabrielfalcao/sure/compare/1.2.9...v1.3.0
 [1.2.9]: https://github.com/gabrielfalcao/sure/compare/1.2.5...1.2.9